2508042127-add- 입장 시점에 따른 사용자별 채팅 메세지 조회 쿼리 수정 및 1:1 채팅 중 사용자 한 명이 나갔…

+++ src/main/java/com/example/grouvy/chat/vo/ChatMessage.java
@@ -14,7 +14,7 @@ import org.apache.ibatis.type.Alias;
 public class ChatMessage {
 
   private long chatMessageId;
-  private long roomId;
+  private int roomId;
   private int senderId;
   private String content;
   private Date createdDate;
